requirements.txt
pypresence
PyGetWindow
pywin32
pystray
pillow
pyinstaller
requests
packaging